requirements.txt
anthropic
google-cloud-aiplatform
gitpython
magika
streamlit
vertexai